+++ model/web/realisasianggaran_update_request.go
@@ -6,7 +6,7 @@ type RealisasiAnggaranUpdateRequest struct {
 	IdPohon                    int    `json:"id_pohon" validate:"required"`
 	IdRencanaKinerja           string `json:"id_rencana_kinerja" validate:"required"`
 	KodeSubkegiatan            string `json:"kode_subkegiatan" validate:"required"`
-	RealisasiAnggaran          int    `json:"realisasi_anggaran" validate:"required"`
+	RealisasiAnggaran          int64  `json:"realisasi_anggaran" validate:"required"`
 	KodeOpd                    string `json:"kode_opd" validate:"required"`
 	RencanaAksi                string `json:"rencana_aksi"`
 	FaktorPendorong            string `json:"faktor_pendorong"`
